Move blocs to `app/lib/blackboard`.
