+++ app/lib/blackboard/blocs/blackboard_dialog_bloc.dart
