+++ app/lib/dashboard/widgets/blackboard_card_dashboard.dart
@@ -8,8 +8,8 @@
 
 import 'package:bloc_provider/bloc_provider.dart';
 import 'package:flutter/material.dart';
+import 'package:sharezone/blackboard/blocs/blackboard_card_bloc.dart';
 import 'package:sharezone/blocs/application_bloc.dart';
-import 'package:sharezone/blocs/blackboard/blackboard_card_bloc.dart';
 import 'package:sharezone/widgets/blackboard/blackboard_card.dart';
 import 'package:sharezone/widgets/blackboard/blackboard_view.dart';
 import 'package:sharezone_common/helper_functions.dart';
